Answer the lanes the first pass left red
The build's own corrections, none of them small enough to miss twice:

- Seven statements in the import tests closed a paren the expression did
  not open; the parse error stood in front of everything behind it.
- The dock's dismiss moved the card's own id out of a children closure
  that re-runs; the swatch grid's per-render clone is the spelling that
  keeps the closure Fn.
- The drop hit-test cast its f64s down to elementFromPoint's f32s; the
  selection's arrivals are collected before the extend so `placed`'s
  borrow of `ordered` ends with the filter; a replace's walk takes a
  clone of the root the gate is still borrowing.
- BookFileRequest declares from before id, the order both shapes it
  replaces already spelled, so a relocation's wire bytes are unchanged
  by the unification.
- A drive root ("C:\\") is all root: file_name answers empty for it,
  so the label falls back to the whole path.
- bind_returned binds by the ADDRESS the log and the row share: the
  row's fp is its own copy's stamp and the log's is the file's, so the
  fingerprints can never meet again — the last_path the log remembers
  and the source the row carries are the one identity the copy cannot
  change. Name stays the fallback for a pending row with no address.
- The ask's action keeps its count-free wording: "2 Move shelves" is
  not a sentence either, so doing() is back beside the counting
  plural() it is not.
- ARCHITECTURE.md names BookFileRequest, the re-export list drops the
  add_link nothing referred to, note.rs drops a prelude it outgrew,
  and a thread_local wears a plain comment — there is no item for a
  doc comment to attach to there.

+++ src/features/library/dnd/target.rs
@@ -74,7 +74,7 @@ impl DropTargetRegistry {
     /// card is a DOM descendant of the level, so the pointer's deepest element reaches it
     /// first — and among targets sharing one node the reverse order still decides.
     pub fn hit_test(&self, x: f64, y: f64) -> Option<DropTargetId> {
-        let hit = web_sys::window()?.document()?.element_from_point(x, y)?;
+        let hit = web_sys::window()?.document()?.element_from_point(x as f32, y as f32)?;
         self.entries.with_untracked(|list| {
             let mut node: Option<web_sys::Element> = Some(hit);
             while let Some(el) = node {
